refactor(crypto): keep bundle size unchanged
